Defined aliases for commands which exist. Made alias names where possible.

+++ .config/zsh/aliases.zsh
@@ -1,10 +1,12 @@
 # common aliases
 
-alias bd='brew update'
-alias bg='brew upgrade'
-alias bi='brew install'
-alias br='brew remove'
-alias bs='brew search'
+if [[ "$(which brew)" == *"brew" ]]; then
+	alias bd='brew update'
+	alias bg='brew upgrade'
+	alias bi='brew install'
+	alias br='brew remove'
+	alias bs='brew search'
+fi
 
 [[ -f "$(which eza)" ]] && alias ls=eza
 
@@ -19,15 +21,17 @@ if [[ "$OSTYPE" == "linux-gnu"* ]]; then
 	# Linux distribution specific aliases
 	export LSB_DISTRIBUTOR=$(lsb_release -i)
 	if [[ "$LSB_DISTRIBUTOR" == *"Fedora"* ]]; then
-		alias sdc="dnf check-update"
-		alias sdu="sudo dnf upgrade"
-		alias sdi="sudo dnf install"
-		alias sdr="sudo dnf remove"
+		alias dc="dnf check-update"
+		alias du="sudo dnf upgrade"
+		alias di="sudo dnf install"
+		alias dr="sudo dnf remove"
+		alias sdr="dnf search"
 	elif [[ "$LSB_DISTRIBUTOR" == *"Ubuntu"* ]]; then
-		alias sac="sudo apt update"
-		alias sau="sudo apt upgrade"
-		alias sai="sudo apt install"
-		alias sar="sudo apt remove"
+		alias ad="sudo apt update"
+		alias au="sudo apt upgrade"
+		alias ai="sudo apt install"
+		alias ar="sudo apt remove"
+		alias as="apt search"
 	fi
 elif [[ "$OSTYPE" == "darwin"* ]]; then
 	[[ -d "/Applications/VisIt.app" ]] && alias visit="/Applications/VisIt.app/Contents/Resources/bin/visit"
